pyproject.toml
[project]
name = "generative_policy_proposals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "torch==2.7.1",
    "huggingface-hub>=0.34.3",
    "peft>=0.16.0",
    #"flash-attn>=2.8.2",
    "decord>=0.6.0",
    "transformers",                                                                                             #==4.51.0",
    "ale-py>=0.11.2",
    "gymnasium[other]>=1.2.0",
    "av>=15.0.0",
    "qwen-vl-utils[decord]>=0.0.8",
    "torchvision==0.22.1",
    "vllm==0.10.0",
    "xformers @ git+https://github.com/facebookresearch/xformers.git@fde5a2fb46e3f83d73e2974a4d12caf526a4203e",
    "blake3>=1.0.5",
    "numba>=0.61.2",
    "einops>=0.8.1",
    "accelerate>=1.9.0",
    "python-slugify>=8.0.4",
]

[tool.uv]
environments = ["sys_platform == 'linux' and platform_machine=='x86_64'"]
prerelease = "if-necessary-or-explicit"
index-strategy = "unsafe-best-match"
package = true

[tool.uv.sources]
torch = [{ index = "pytorch-stable-cu128" }]
torchvision = [{ index = "pytorch-stable-cu128" }]

[[tool.uv.index]]
name = "pytorch-stable-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = false

[[tool.uv.dependency-metadata]]
name = "vllm"
version = "0.10.0"
requires-dist = [
    "regex",                                                                                                                             # Replace re for higher-performance regex matching
    "cachetools",
    "psutil",
    "sentencepiece",                                                                                                                     # Required for LLaMA tokenizer.
    "numpy",
    "requests >= 2.26.0",
    "tqdm",
    "blake3",
    "py-cpuinfo",
    "transformers >= 4.53.2",
    "huggingface-hub[hf_xet] >= 0.33.0",                                                                                                 # Required for Xet downloads.
    "tokenizers >= 0.21.1",                                                                                                              # Required for fast incremental detokenization.
    "protobuf",                                                                                                                          # Required by LlamaTokenizer.
    "fastapi[standard] >= 0.115.0",                                                                                                      # Required by FastAPI's form models in the OpenAI API server's audio transcriptions endpoint.
    "aiohttp",
    "openai >= 1.87.0",                                                                                                                  # Ensure modern openai package (ensure ResponsePrompt exists in type.responses and max_completion_tokens field support)
    "pydantic >= 2.10",
    "prometheus_client >= 0.18.0",
    "pillow",                                                                                                                            # Required for image processing
    "prometheus-fastapi-instrumentator >= 7.0.0",
    "tiktoken >= 0.6.0",                                                                                                                 # Required for DBRX tokenizer
    "lm-format-enforcer >= 0.10.11, < 0.11",
    "llguidance >= 0.7.11, < 0.8.0; platform_machine == \"x86_64\" or platform_machine == \"arm64\" or platform_machine == \"aarch64\"",
    "outlines_core == 0.2.10",
    # required for outlines backend disk cache
    "diskcache == 5.6.3",
    "lark == 1.2.2",
    "xgrammar == 0.1.21; platform_machine == \"x86_64\" or platform_machine == \"aarch64\" or platform_machine == \"arm64\"",
    "typing_extensions >= 4.10",
    "filelock >= 3.16.1",                                                                                                     # need to contain https://github.com/tox-dev/filelock/pull/317
    "partial-json-parser",                                                                                                    # used for parsing partial JSON outputs
    "pyzmq >= 25.0.0",
    "msgspec",
    "gguf >= 0.13.0",
    "importlib_metadata; python_version < '3.10'",
    "mistral_common[image,audio] >= 1.8.2",
    "opencv-python-headless >= 4.11.0",                                                                                       # required for video IO
    "pyyaml",
    "six>=1.16.0; python_version > '3.11'",                                                                                   # transitive dependency of pandas that needs to be the latest version for python 3.12
    "einops",                                                                                                                 # Required for Qwen2-VL.
    "compressed-tensors == 0.10.2",                                                                                           # required for compressed-tensors
    "depyf==0.19.0",                                                                                                          # required for profiling and debugging with compilation config
    "cloudpickle",                                                                                                            # allows pickling lambda functions in model_executor/models/registry.py
    "watchfiles",                                                                                                             # required for http server to monitor the updates of TLS files
    "python-json-logger",                                                                                                     # Used by logging as per examples/others/logging_configuration.md
    "scipy",                                                                                                                  # Required for phi-4-multimodal-instruct
    "ninja",                                                                                                                  # Required for xgrammar, rocm, tpu, xpu
    "pybase64",                                                                                                               # fast base64 implementation
    "cbor2",                                                                                                                  # Required for cross-language serialization of hashable objects
    "setproctitle",                                                                                                           # Used to set process names for better debugging and monitoring
    "xformers @ git+https://github.com/facebookresearch/xformers.git@fde5a2fb46e3f83d73e2974a4d12caf526a4203e",
]
